hide User() and WebClient()
