+++ shepherd_core/examples/firmware_model.py
@@ -8,8 +8,8 @@ from pathlib import Path
 
 from shepherd_core.data_models import Firmware
 from shepherd_core.data_models import FirmwareDType
+from shepherd_core.testbed_client.client_web import WebClient
 
-from shepherd_core import WebClient
 from shepherd_core import fw_tools
 
 path_elf = Path(__file__).parent.parent / "tests/fw_tools/build_msp.elf"
