+++ shepherd_core/shepherd_core/__init__.py
@@ -19,7 +19,6 @@ from .logger import get_verbose_level
 from .logger import increase_verbose_level
 from .logger import log
 from .reader import Reader
-from .testbed_client.client_web import WebClient
 from .version import version
 from .writer import Writer
 
@@ -35,7 +34,6 @@ __all__ = [
     "Compression",
     "Inventory",
     "Reader",
-    "WebClient",
     "Writer",
     "get_verbose_level",
     "increase_verbose_level",
